Test CLI version command
