--- test/advent/configuration_test.rb
@@ -1,4 +1,0 @@
-require "test_helper"
-
-class Advent::ConfigurationTest < Advent::TestCase
-end
