--- test/advent/input_test.rb
@@ -1,4 +1,0 @@
-require "test_helper"
-
-class Advent::InputTest < Advent::TestCase
-end
