--- test/advent/session_test.rb
@@ -1,4 +1,0 @@
-require "test_helper"
-
-class Advent::SessionTest < Advent::TestCase
-end
