--- test/advent/solution_test.rb
@@ -1,4 +1,0 @@
-require "test_helper"
-
-class Advent::SolutionTest < Advent::TestCase
-end
